chore(api): rearranging routes into folders
