+++ apps/api/src/routes/v1/proposals/_id/votes/schema.ts
@@ -1,6 +1,18 @@
 import { Type, Static } from 'typebox';
 import { userVote } from '@schemas/proposal';
 
+export const params = Type.Object({
+  id: Type.Integer({
+    minimum: 0,
+    description: 'Proposal ID',
+  }),
+});
+
+export const query = Type.Object({
+  limit: Type.Integer({ default: 20, minimum: 0, maximum: 100 }),
+  offset: Type.Integer({ default: 0, minimum: 0 }),
+});
+
 const voter = Type.Object({
   name: Type.String(),
   surname: Type.String(),
@@ -19,4 +31,14 @@ export const response = Type.Object({
   }),
 });
 
-export type Response = Static<typeof response>;
+export const schema = {
+  params,
+  querystring: query,
+  response: { 200: response },
+} as const;
+
+export type Schema = {
+  Reply: Static<typeof response>;
+  Params: Static<typeof params>;
+  Querystring: Static<typeof query>;
+};
