--- apps/api/src/routes/v1/proposals/schema/get-proposals-id-voters/index.ts
@@ -1,3 +1,0 @@
-export * from './params';
-export * from './query';
-export * from './response';
