--- apps/api/src/routes/v1/proposals/schema/get-proposals-id-voters/params.ts
@@ -1,10 +1,0 @@
-import { Type, Static } from 'typebox';
-
-export const params = Type.Object({
-  id: Type.Integer({
-    minimum: 0,
-    description: 'Proposal ID',
-  }),
-});
-
-export type Params = Static<typeof params>;
