--- apps/api/src/routes/v1/proposals/schema/get-proposals-id-voters/query.ts
@@ -1,8 +1,0 @@
-import { Type, Static } from 'typebox';
-
-export const query = Type.Object({
-  limit: Type.Integer({ default: 20, minimum: 0, maximum: 100 }),
-  offset: Type.Integer({ default: 0, minimum: 0 }),
-});
-
-export type Query = Static<typeof query>;
