--- apps/api/src/routes/v1/proposals/schema/get-proposals-id/index.ts
@@ -1,2 +1,0 @@
-export * from './params';
-export * from './response';
