--- apps/api/src/routes/v1/proposals/schema/get-proposals-id/response.ts
@@ -1,10 +1,0 @@
-import { Type, Static } from 'typebox';
-import { summary, commitment, payment } from '@schemas/proposal';
-
-export const response = Type.Object({
-  ...summary.properties,
-  commitment: Type.Optional(commitment),
-  payments: Type.Array(payment),
-});
-
-export type Response = Static<typeof response>;
