--- apps/api/src/routes/v1/proposals/schema/get-proposals/index.ts
@@ -1,2 +1,0 @@
-export * from './query';
-export * from './response';
