--- apps/api/src/routes/v1/proposals/schema/get-proposals/response.ts
@@ -1,13 +1,0 @@
-import { Type, Static } from 'typebox';
-import { summary } from '@schemas/proposal';
-
-export const response = Type.Object({
-  data: Type.Array(summary),
-  meta: Type.Object({
-    total: Type.Integer(),
-    limit: Type.Integer(),
-    offset: Type.Integer(),
-  }),
-});
-
-export type Response = Static<typeof response>;
